+++ appIcon.h
@@ -34,7 +34,7 @@ void appIcon_ReplyDialog(int mPos, char* qText);
 
 void appIcon_UpdateLists(struct NODE *llist, struct Broadcast_Packet* p_Item, char m_Option);
 
-Boolean appIcon_SearchList(int* pNum, char* strPacket);
+Boolean appIcon_SearchList(long* pNum, char* strPacket);
 Boolean appIcon_SearchNode(struct SendClientData* num, int* mIdx, char* strHostname);
 Boolean appIcon_SearchItems(struct SendClientData* num, int* mIdx, char* strFileID);
